RAG_based_QA_system/requirements.txt
Flask
langchain
langchain-community
faiss-cpu
PyMuPDF
python-dotenv
ollama
